requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
langchain>=0.1.0
langchain-core>=0.1.30
langchain-community>=0.0.24
langchain-google-genai>=0.0.11
faiss-cpu>=1.7.4